fix: ensure shared selection helpers import
