+++ tests/unit/_support/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './keyboard';
 export * from './note';
+export * from './selection';
