+++ tests/unit/selection.spec.ts
@@ -1,11 +1,14 @@
 import { act, waitFor, within } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { placeCaretAtNote, pressKey, readOutline } from '#tests';
 import type { Outline } from '#tests';
+import {
+  collectSelectedListItems,
+  getListItemLabel,
+  placeCaretAtNote,
+  pressKey,
+  readOutline,
+} from '#tests';
 import { $getSelection, $isRangeSelection } from 'lexical';
-import type { LexicalNode, RangeSelection } from 'lexical';
-import { $isListItemNode, $isListNode } from '@lexical/list';
-import type { ListItemNode } from '@lexical/list';
 
 const TREE_COMPLEX_OUTLINE: Outline = [
   {
@@ -18,78 +21,6 @@ const TREE_COMPLEX_OUTLINE: Outline = [
   { text: 'note5', children: [] },
   { text: 'note6', children: [ { text: 'note7', children: [] } ] },
 ];
-
-function collectSelectedListItems(selection: RangeSelection): ListItemNode[] {
-  const seen = new Set<string>();
-  const items: ListItemNode[] = [];
-
-  for (const node of selection.getNodes()) {
-    const listItem = findNearestListItem(node);
-    if (!listItem || !listItem.isAttached()) {
-      continue;
-    }
-
-    const key = listItem.getKey();
-    if (seen.has(key)) {
-      continue;
-    }
-
-    seen.add(key);
-    items.push(listItem);
-  }
-
-  return items.toSorted((a, b) => (a === b ? 0 : a.isBefore(b) ? -1 : 1));
-}
-
-function getListItemLabel(item: ListItemNode): string | null {
-  const contentItem = resolveContentListItem(item);
-  const pieces: string[] = [];
-  for (const child of contentItem.getChildren()) {
-    if (typeof child.getType === 'function' && child.getType() === 'list') {
-      continue;
-    }
-
-    const getTextContent = (child as { getTextContent?: () => string }).getTextContent;
-    if (typeof getTextContent === 'function') {
-      pieces.push(getTextContent.call(child));
-    }
-  }
-
-  const label = pieces.join('').trim();
-  if (label.length > 0) {
-    return label;
-  }
-
-  return contentItem === item ? null : getListItemLabel(contentItem);
-}
-
-function resolveContentListItem(item: ListItemNode): ListItemNode {
-  if (!isChildrenWrapper(item)) {
-    return item;
-  }
-
-  const previous = item.getPreviousSibling();
-  return $isListItemNode(previous) ? previous : item;
-}
-
-function isChildrenWrapper(node: LexicalNode | null): boolean {
-  if (!$isListItemNode(node)) {
-    return false;
-  }
-  const children = node.getChildren();
-  return children.length === 1 && $isListNode(children[0] ?? null);
-}
-
-function findNearestListItem(node: LexicalNode | null): ListItemNode | null {
-  let current: LexicalNode | null = node;
-  while (current) {
-    if ($isListItemNode(current)) {
-      return resolveContentListItem(current);
-    }
-    current = current.getParent();
-  }
-  return null;
-}
 
 // Ensures every multi-note selection matches the guarantees from docs/selection.md:
 // once a selection crosses a note boundary it must cover a contiguous block of
